MAJOR update , refactored the entire comparinator and over engineered the shit out of it. It compares, it inspects, it provides you with the parameters to remake it , it saves, fuck it does everything. It's a one stop shop for preview, save, compare. Added a translation node I'm working on cause I have a few Chinese nodes. Updated Gridinator to no longer fuck up on refresh, updated model save to be oom friendly- Should never OoM now, updated Merger to iteratively merge and calculate the weights based on the values set by user and the direction of the merge. I.e. Merges better but a little slower, plus side , no more ''unable to determine model type'' and no more ''can't find location _SD''. Like I said Major update.

+++ gguf/dequant.py
@@ -24,7 +24,7 @@ def dequantize_tensor(tensor, dtype=None, dequant_dtype=None):
     else:
         # this is incredibly slow
         tqdm.write(f"Falling back to numpy dequant for qtype: {getattr(qtype, 'name', repr(qtype))}")
-        new = gguf.quants.dequantize(tensor.cpu().numpy(), qtype)
+        new = gguf.quants.dequantize(tensor.cpu().float().numpy(), qtype)
         return torch.from_numpy(new).to(tensor.device, dtype=dtype)
 
 def dequantize(data, qtype, oshape, dtype=None):
